host-requirements.txt
psutil
asyncio
python-socketio
aiohttp